api: split types.go into multiple distinct type specific files
This helps break up the humongous types.go file into more distinct, type-specific
files. There are also minor renaming changes here, where a C_STYLE_CONST was
renamed as a GoStyleConst and comments have been added as per Lint guidelines
