+++ agent/api/ecr.go
@@ -1,0 +1,28 @@
+// Copyright 2014-2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package api
+
+// RegistryAuthenticationData is the authentication data sent by the ECS backend.  Currently, the only supported
+// authentication data is for ECR.
+type RegistryAuthenticationData struct {
+	Type        string       `json:"type"`
+	ECRAuthData *ECRAuthData `json:"ecrAuthData"`
+}
+
+// ECRAuthData is the authentication details for ECR specifying the region, registryID, and possible endpoint override
+type ECRAuthData struct {
+	EndpointOverride string `json:"endpointOverride"`
+	Region           string `json:"region"`
+	RegistryID       string `json:"registryId"`
+}
