+++ agent/api/task.go
@@ -16,9 +16,11 @@ package api
 import (
 	"encoding/json"
 	"errors"
+	"fmt"
 	"path/filepath"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/aws/amazon-ecs-agent/agent/acs/model/ecsacs"
@@ -38,6 +40,70 @@ const (
 	// credentials.
 	awsSDKCredentialsRelativeURIPathEnvironmentVariableName = "AWS_CONTAINER_CREDENTIALS_RELATIVE_URI"
 )
+
+// TaskOverrides are the overrides applied to a task
+type TaskOverrides struct{}
+
+// TaskVolume is a definition of all the volumes available for containers to
+// reference within a task. It must be named.
+type TaskVolume struct {
+	Name   string `json:"name"`
+	Volume HostVolume
+}
+
+// Task is the internal representation of a task in the ECS agent
+type Task struct {
+	// Arn is the unique identifer for the task
+	Arn string
+	// Overrides are the overrides applied to a task
+	Overrides TaskOverrides `json:"-"`
+	// Family is the name of the task definition family
+	Family string
+	// Version is the version of the task definition
+	Version string
+	// Containers are the containers for the task
+	Containers []*Container
+	// Volumes are the volumes for the task
+	Volumes []TaskVolume `json:"volumes"`
+
+	// DesiredStatus represents the state where the task should go.  Generally the desired status is informed by the ECS
+	// backend as a result of either API calls made to ECS or decisions made by the ECS service scheduler.  The
+	// DesiredStatus is almost always either TaskRunning or TaskStopped.  Do not access DesiredStatus directly.  Instead,
+	// use `UpdateStatus`, `UpdateDesiredStatus`, `SetDesiredStatus`, and `SetDesiredStatus`.
+	// TODO DesiredStatus should probably be private with appropriately written setter/getter.  When this is done, we need
+	// to ensure that the UnmarshalJSON is handled properly so that the state storage continues to work.
+	DesiredStatus     TaskStatus
+	desiredStatusLock sync.RWMutex
+
+	// KnownStatus represents the state where the task is.  This is generally the minimum of equivalent status types for
+	// the containers in the task; if one container is at ContainerRunning and another is at ContainerPulled, the task
+	// KnownStatus would be TaskPulled.  Do not access KnownStatus directly.  Instead, use `UpdateStatus`,
+	// `UpdateKnownStatusAndTime`,  and `GetKnownStatus`.
+	// TODO KnownStatus should probably be private with appropriately written setter/getter.  When this is done, we need
+	// to ensure that the UnmarshalJSON is handled properly so that the state storage continues to work.
+	KnownStatus     TaskStatus
+	knownStatusLock sync.RWMutex
+	// KnownStatusTime captures the time when the KnownStatus was last updated.  Do not access KnownStatusTime directly,
+	// instead use `GetKnownStatusTime`.
+	KnownStatusTime     time.Time `json:"KnownTime"`
+	knownStatusTimeLock sync.RWMutex
+
+	// SentStatus represents the last KnownStatus that was sent to the ECS SubmitTaskStateChange API.
+	// TODO(samuelkarp) SentStatus needs a lock and setters/getters.
+	// TODO SentStatus should probably be private with appropriately written setter/getter.  When this is done, we need
+	// to ensure that the UnmarshalJSON is handled properly so that the state storage continues to work.
+	SentStatus     TaskStatus
+	sentStatusLock sync.RWMutex
+
+	StartSequenceNumber int64
+	StopSequenceNumber  int64
+
+	// credentialsID is used to set the CredentialsId field for the
+	// IAMRoleCredentials object associated with the task. This id can be
+	// used to look up the credentials for task in the credentials manager
+	credentialsID     string
+	credentialsIDLock sync.RWMutex
+}
 
 // PostUnmarshalTask is run after a task has been unmarshalled, but before it has been
 // run. It is possible it will be subsequently called after that and should be
@@ -99,7 +165,7 @@ func (task *Task) initializeEmptyVolumes() {
 
 // initializeCredentialsEndpoint sets the credentials endpoint for all containers in a task if needed.
 func (task *Task) initializeCredentialsEndpoint(credentialsManager credentials.Manager) {
-	id := task.GetCredentialsId()
+	id := task.GetCredentialsID()
 	if id == "" {
 		// No credentials set for the task. Do not inject the endpoint environment variable.
 		return
@@ -165,17 +231,6 @@ func (task *Task) UpdateMountPoints(cont *Container, vols map[string]string) {
 					empty.HostPath = hostPath
 				}
 			}
-		}
-	}
-}
-
-// updateContainerDesiredStatus sets all container's desired status's to the
-// task's desired status
-func (task *Task) updateContainerDesiredStatus() {
-	for _, c := range task.Containers {
-		taskDesiredStatus := task.GetDesiredStatus()
-		if c.GetDesiredStatus() < taskDesiredStatus.ContainerStatus() {
-			c.SetDesiredStatus(taskDesiredStatus.ContainerStatus())
 		}
 	}
 }
@@ -248,8 +303,8 @@ func (task *Task) dockerConfig(container *Container) (*docker.Config, *DockerCli
 
 	// Convert MB to B
 	dockerMem := int64(container.Memory * 1024 * 1024)
-	if dockerMem != 0 && dockerMem < DOCKER_MINIMUM_MEMORY {
-		dockerMem = DOCKER_MINIMUM_MEMORY
+	if dockerMem != 0 && dockerMem < DockerContainerMinimumMemoryInBytes {
+		dockerMem = DockerContainerMinimumMemoryInBytes
 	}
 
 	var entryPoint []string
@@ -265,7 +320,7 @@ func (task *Task) dockerConfig(container *Container) (*docker.Config, *DockerCli
 		Volumes:      dockerVolumes,
 		Env:          dockerEnv,
 		Memory:       dockerMem,
-		CPUShares:    task.dockerCpuShares(container.CPU),
+		CPUShares:    task.dockerCPUShares(container.CPU),
 	}
 
 	if container.DockerConfig.Config != nil {
@@ -281,16 +336,16 @@ func (task *Task) dockerConfig(container *Container) (*docker.Config, *DockerCli
 	return config, nil
 }
 
+// dockerCPUShares converts containerCPU shares if needed as per the logic stated below:
 // Docker silently converts 0 to 1024 CPU shares, which is probably not what we
 // want.  Instead, we convert 0 to 2 to be closer to expected behavior. The
-// reason for 2 over 1 is that 1 is an invalid value (Linux's choice, not
-// Docker's).
-func (task *Task) dockerCpuShares(containerCpu uint) int64 {
-	if containerCpu <= 1 {
-		log.Debug("Converting CPU shares to allowed minimum of 2", "task", task.Arn, "cpuShares", containerCpu)
+// reason for 2 over 1 is that 1 is an invalid value (Linux's choice, not Docker's).
+func (task *Task) dockerCPUShares(containerCPU uint) int64 {
+	if containerCPU <= 1 {
+		log.Debug("Converting CPU shares to allowed minimum of 2", "task", task.Arn, "cpuShares", containerCPU)
 		return 2
 	}
-	return int64(containerCpu)
+	return int64(containerCPU)
 }
 
 func (task *Task) dockerExposedPorts(container *Container) map[docker.Port]struct{} {
@@ -472,6 +527,22 @@ func TaskFromACS(acsTask *ecsacs.Task, envelope *ecsacs.PayloadMessage) (*Task, 
 	return task, nil
 }
 
+// UpdateStatus updates a task's known and desired statuses to be compatible
+// with all of its containers
+// It will return a bool indicating if there was a change
+func (task *Task) UpdateStatus() bool {
+	change := task.updateTaskKnownStatus()
+	// DesiredStatus can change based on a new known status
+	task.UpdateDesiredStatus()
+	return change != TaskStatusNone
+}
+
+// SetKnownStatus sets the known status of the task
+func (task *Task) UpdateDesiredStatus() {
+	task.updateTaskDesiredStatus()
+	task.updateContainerDesiredStatus()
+}
+
 // updateTaskDesiredStatus determines what status the task should properly be at based on its container's statuses
 func (task *Task) updateTaskDesiredStatus() {
 	llog := log.New("task", task)
@@ -487,24 +558,35 @@ func (task *Task) updateTaskDesiredStatus() {
 	}
 }
 
-// UpdateStatus updates a task's known and desired statuses to be compatible
-// with all of its containers
-// It will return a bool indicating if there was a change
-func (task *Task) UpdateStatus() bool {
-	change := task.updateTaskKnownStatus()
-	// DesiredStatus can change based on a new known status
-	task.UpdateDesiredStatus()
-	return change != TaskStatusNone
+// updateContainerDesiredStatus sets all container's desired status's to the
+// task's desired status
+func (task *Task) updateContainerDesiredStatus() {
+	for _, c := range task.Containers {
+		taskDesiredStatus := task.GetDesiredStatus()
+		if c.GetDesiredStatus() < taskDesiredStatus.ContainerStatus() {
+			c.SetDesiredStatus(taskDesiredStatus.ContainerStatus())
+		}
+	}
 }
 
-func (task *Task) UpdateDesiredStatus() {
-	task.updateTaskDesiredStatus()
-	task.updateContainerDesiredStatus()
-}
-
+// SetKnownStatus sets the known status of the task
 func (task *Task) SetKnownStatus(status TaskStatus) {
 	task.setKnownStatus(status)
 	task.updateKnownStatusTime()
+}
+
+func (task *Task) setKnownStatus(status TaskStatus) {
+	task.knownStatusLock.Lock()
+	defer task.knownStatusLock.Unlock()
+
+	task.KnownStatus = status
+}
+
+func (task *Task) updateKnownStatusTime() {
+	task.knownStatusTimeLock.Lock()
+	defer task.knownStatusTimeLock.Unlock()
+
+	task.KnownStatusTime = ttime.Now()
 }
 
 // UpdateKnownStatusAndTime updates the KnownStatus and KnownStatusTime
@@ -530,34 +612,23 @@ func (task *Task) GetKnownStatusTime() time.Time {
 	return task.KnownStatusTime
 }
 
-func (task *Task) setKnownStatus(status TaskStatus) {
-	task.knownStatusLock.Lock()
-	defer task.knownStatusLock.Unlock()
-
-	task.KnownStatus = status
-}
-
-func (task *Task) updateKnownStatusTime() {
-	task.knownStatusTimeLock.Lock()
-	defer task.knownStatusTimeLock.Unlock()
-
-	task.KnownStatusTime = ttime.Now()
-}
-
-func (task *Task) SetCredentialsId(id string) {
+// SetCredentialsID sets the credentials ID for the task
+func (task *Task) SetCredentialsID(id string) {
 	task.credentialsIDLock.Lock()
 	defer task.credentialsIDLock.Unlock()
 
 	task.credentialsID = id
 }
 
-func (task *Task) GetCredentialsId() string {
+// GetCredentialsID gets the credentials ID for the task
+func (task *Task) GetCredentialsID() string {
 	task.credentialsIDLock.RLock()
 	defer task.credentialsIDLock.RUnlock()
 
 	return task.credentialsID
 }
 
+// GetDesiredStatus gets the desired status of the task
 func (task *Task) GetDesiredStatus() TaskStatus {
 	task.desiredStatusLock.RLock()
 	defer task.desiredStatusLock.RUnlock()
@@ -565,6 +636,7 @@ func (task *Task) GetDesiredStatus() TaskStatus {
 	return task.DesiredStatus
 }
 
+// SetDesiredStatus sets the desired status of the task
 func (task *Task) SetDesiredStatus(status TaskStatus) {
 	task.desiredStatusLock.Lock()
 	defer task.desiredStatusLock.Unlock()
@@ -586,4 +658,14 @@ func (task *Task) SetSentStatus(status TaskStatus) {
 	defer task.sentStatusLock.Unlock()
 
 	task.SentStatus = status
+}
+
+// String returns a human readable string representation of this object
+func (t *Task) String() string {
+	res := fmt.Sprintf("%s:%s %s, Status: (%s->%s)", t.Family, t.Version, t.Arn, t.GetKnownStatus().String(), t.GetDesiredStatus().String())
+	res += " Containers: ["
+	for _, c := range t.Containers {
+		res += fmt.Sprintf("%s (%s->%s),", c.Name, c.GetKnownStatus().String(), c.GetDesiredStatus().String())
+	}
+	return res + "]"
 }
